Adicionando versão vanila 0.1.0

+++ src/main/java/org/vanilaGraph/Shortcut.java
@@ -4,7 +4,7 @@ public class Shortcut extends Edge{
 	
 	private int contractedNodeIndex;
 
-	public Shortcut(Long edgeId, int fromIndex, int toIndex, Double cost, int contractedNodeIndex) {
+	public Shortcut(Long edgeId, int fromIndex, int toIndex, int cost, int contractedNodeIndex) {
 		super(edgeId, fromIndex, toIndex, cost);
 		this.contractedNodeIndex = contractedNodeIndex;
 	}
